reverted to rev 69. stupid changes.
the database field timestamp is now a java field an gets initialized
by object construction. this allows the user to get the timestamp, which
would otherwise be null (this is also why getTimeStamp() was stupid before).
This was changed to allow for a consistent user experience (if you can specify
timestamps, you should be able to read them in your results).

It was previously a database field, which was initialized by default by the database,
if an object of that type is persisted. that way, the timestamp is always correct.
now you can hold your reference to your AccountancyEntry and wait a potentially long
time before persisting the entry - as a result, the timestamp will be off.



git-svn-id: svn://141.76.65.164/kp-salespoint@99 eb53f5e8-9ec5-4bc0-8d13-06c033283886