adds proper annotation to AbstractServiceInstance.
AbstractProductInstance has an ID-field in its embedded PK.
So does ServiceType (AbstractProductType).
This will result in two ID-fields for the same table. One of
those fields has to be declared read-only, or a proper
@JoinColumn has to be declared.
See:
http://en.wikibooks.org/wiki/Java_Persistence/OneToOne#Advanced
or ECLIPSELINK-00048 in 
http://wiki.eclipse.org/EclipseLink_Exception_Error_Reference_(ELUG)

PS: This this patch /never/, /never ever/ worked. Please do /NOT/
commit broken code (again).


git-svn-id: svn://141.76.65.164/kp-salespoint@279 eb53f5e8-9ec5-4bc0-8d13-06c033283886