adds Dummy classes to allow project build despite abstract entities
Again, this is clearly a bug in eclipselink, as JPA-spec states
abstract classes can be entities. However, we have to deal with it.
A proper solution would be to implement our own SessionCustomizer,
to work around this problem. This is, however, not easy.
Maybe these classes should be distributed in sp.jar, so that a blank
project can be built.


git-svn-id: svn://141.76.65.164/kp-salespoint@280 eb53f5e8-9ec5-4bc0-8d13-06c033283886