adds requested Accountancy features
methods to query for entries with time spec.
spec include from, to and period.
the time between from and to is divided into fragments of time
of period length. for each fragment, 
a) an iterable is returned, containing all entries within that fragment
b) a money object is returned representing the sum of the value of all
entries within that fragment.

this allows you to easily create charts etc; for example
plot monthly salesvolume over the last two years:
a.getSalesVolume(ProductPaymentEntry.class, new DateTime().minusYears(2), new DateTime(), Period.months(1));
this will return an Iterable<Interval, Money> containing the sales volume in Money
for every month, specified in Interval.



git-svn-id: svn://141.76.65.164/kp-salespoint@326 eb53f5e8-9ec5-4bc0-8d13-06c033283886