Introduced AuthenticationManager component to ease access to currently logged in user. (Oliver Gierke)
The newly introduced AuthenticationManager interface provides access to the currently logged in UserAccount and will be available as a Spring bean in the ApplicationContext. Added a Spring Security based implementation to avoid clients tying to Spring Security and static factory methods. Changed the implementation of LoggedInUserAccountArgumentResolver to use the AuthenticationManager component. Polished test cases.

git-svn-id: svn://141.76.65.164/kp-salespoint@1051 eb53f5e8-9ec5-4bc0-8d13-06c033283886