Further refinement to configuration setup - (Oliver Gierke)
Pulled up more default security configuration into Salespoint to reduce the amount of configuration necessary in student projects. Exclude @Configuration classes from default component scanning to be able to selectively use configuration classes in student projects. This is necessary as the projects have to extend the SalespointSecurityConfiguration class and *not* include it as is. Updates to the sample student projects follow with the next commit.

Added ….matches(Password, Password) method to AuthenticationManager to prevent the need to use the SpringSecurity PasswordEncoder directly.

git-svn-id: svn://141.76.65.164/kp-salespoint@1063 eb53f5e8-9ec5-4bc0-8d13-06c033283886