Added DataInitializer abstraction. (Oliver Gierke)
To avoid LazyLoadingExceptions when writing DataInitializers in student projects I introduced a dedicated DataInitializer interface that has a single transactional method. The SalespointDataInitializer will collect all implementations of the interface and invoke the initialize method on the so that the actual initializing code provided by the user will run inside a transaction and thus a single JPA EntityManager Session.

git-svn-id: svn://141.76.65.164/kp-salespoint@1085 eb53f5e8-9ec5-4bc0-8d13-06c033283886