#7, #8 - Overhaul in business time management.
Removed dependency on JodaTime and switched to Java 8's new DateTime APIs. Introduced Interval value object as the latter is missing the concept of an interval with dedicated start and end date(time).

Pulled in Jadira Usertypes to be able to store Java 8 types in Hibernate directly as the conversion from and into legacy Date instances requires dealing with a default time zone which is cumbersome.

Refactored TimeService to BusinessTime and simplified implementation.