#19 - Replaced Catalog implementation with a Spring Data repository.
The Products interface now replaces Catalog. Clients can use the repository either typed to Product or a more concrete domain type. Also, extending the interface to declare additional query methods is a reasonable option as well.

Renamed Keks to Cookie and added a concrete sub-repository to make sure the injection still works if a more concrete repository interface is around.