#17 - Added artifact and site deployment.
Added the necessary plugins and declarations to distribute artifacts, the static website and reference documentation to the Salespoint server. Reconfigured the Asciidoctor and JavaDoc plugin to generate output into folders that are picked up by the site plugin by default.

By default the reference documentation and JavaDocs are generated into a dev folder so that we can continuously deploy website updates alongside current (in-development JavaDocs and reference documentation). The release profile removes that prefix so that release distributions get deployed into the root folder directly.

Updated readme to include instructions to deploy artifacts and the site.