#9 - Introduced simplified Quantity abstraction.
Removed the former Quantity abstraction in favor of a simplified, more easy to understand and pragmatic version. Quantity is now a basic value object and refers to an enum Metric with a fixed set of available metrics.

Mathematical operations on Quantities are only allowed on compatible (i.e. the very same Metric currently) metrics. Attempts to add, subtract or compare Quantities of non-compatible Quantities are rejected with an IllegalArgumentException.

This rather strict arrangement could be softened going forward by introducing other metrics of the same type (e.g. kilometers etc.) and basically adapting the factorial difference within the operation. The code currently uses the concept of compatibility but actually doesn't exercise it.

Adapted all domain objects to use the new Quantity API.

Introduced Lombok as library to ease the implementation of value objects. Tweaked the compiler settings to run Lombok on compilation.