#32 - Added changelog to website.
Added ChangelogCreator class to automatically create a changelog from the solved issues at GitHub. Requires GitHub credentials to be present in an ev.properties file in the project root. The file needs the following two keys:

- github.username
- github.password

Added the env.properties to .gitignore to prevent it from being checked into VCS.

Slightly restructured the Asciidoctor build to be able to build more than one source file for the website. Linked the changelog from the website.