#35 - Added auto-configuration support for sending emails.
When bootstrapping Salespoint an instance of JavaMailSenderImpl is registered as spring bean if the properties spring.mail.host, spring.mail.username and spring.mail.password are present and no bean of type MailSender is defined manually.

Added ConsoleWritingMailSender to replace the default one for testing scenarios so that no real SMTP server is required.