#31 - Yet another round of improvements of the reference documentation.
Removed a lot of references to outdated code. Added a section on the prerequisites (linked the script of the introductory lecture) and one on the Videoshop sample project.

Added UML diagram for user accounts, roles and passwords. Added section on business time API. Added FAQ on how to implement sending emails via Salespoint. Replaced references to "Salespoint 5" with "Salespoint {majorversion}" and exposed {majorversion} through the build.

Some JavaDoc polish here and there.