#34 - Fixed pesistence mapping of InventoryItem.quantity.
Changing the quantity of an InventoryItem caused issues when flushing the changes to the database (org.hsqldb.HsqlException: data exception: string data, right truncation). This can be fixed by manually configuring the length of the lob column (using @Column).

Also, the quantity calculation on decreasing quantities was broken in InventoryItem. Created a test case for that.