#48 - Added findByUsername(…) to UserAccountManager.
To lookup a user by username clients previously had to know that the username is encoded into the UserAccountIdentifier. Thus, looking up a user by username required clients to create new UserAccountIdentifier instances (see VideoShopDataInitializer in 1.0 GA). We added the extra method to prepare the solution of #47.

Polished JavaDoc of UserAccountManager.

Related issues: #47.