#68 - Added unique-constraint to InventoryItem's relation to a Product.
Adding an InventoryItem with a Product which already exists an InventoryItem for now results in a DataIntegrityViolationException (indirectly tested by evaluating the root JPA exception).

This makes sure the query methods declared in Inventory really only return a single instance and a user is forced to lookup a potentially existing InventoryItem for a Product and modifying it.