#113 - Upgraded to Spring Boot 1.3 RC1.
Removed explicit configuration of CharacterEncodingFilter as the bug that required this was fixed in 1.3 RC1. Removed obsolete property files to test mail functionality as this is also part of Spring Boot now. Removed templates placeholder as Boot 1.3 RC1 handles this more gracefully, too.

Removed obsolete MANIFEST.MF and logging config.