#144 - Changed order handling to use events instead of direct method calls.
Refactored PersistentOrderManager to now publish events via the ApplicationEventPublisher, let both Accountancy and Inventory listen to the events and implement the functionality previously implemented in PersistentOrderManager inside the very sub systems.

AccountancyOrderListener now creates the ProductPaymentEntry instances for OrderPaid events and stores them using the according repository.

InventoryOrderListener listens for OrderCompleted events and reduces the stock for all Products referred to by an OrderLine that a LineItemFilter approves. By default, LineItemFilter.handleAll() is configured. Additional LineItemFilter implementations can be registered to customize the reduction process and potentially disable it completely. In case insufficient stock is encountered, the listener throws an OrderCompletionFailed that contains a detailed OrderCompletionReport about which items failed etc.

Updated the reference documentation to include the new way of working with OrderManager. Fixed JavaDoc for OrderManager methods that return a Stremable (the doc was previously pointing to Iterable).

Removed the redundant getIdentifier() method from entities and unify every usage to the already existing getId() method. Moved ProductPaymentEntry into the accounting subsystem. Removed the Apache license header from Streamable and added a method to produce a Streamable from a supplied Stream. Added missing @Transactional annotation to PersistentUserAccountManager.create(…) to make sure we don't run in a read-only transaction (which is the default).