#156 - Fixed parsing side of MonetaryAmountAttributeConverter.
The parsing side of MonetaryAmountAttributeConverter previously assumed to read a formatted string. For monetary amounts with trailing zeros, the writing side of things produces a String using the E notation. As this gets ignored by the formatted read, basically all trailing zeros got los reading the value.

We now try the undformatted read first (to pick up those values properly) falling back to an attempt to read a formatted value to keep the original behavior around in case the database contains formatted values for whatever reason.