#185 - Added formatter for Quantity.
We now ship and register a Spring Formatter for Quantity so that it can immediately be printed and parsed in context of the current Locale and doesn't need to be rendered manually (the amount and metric separately).

The QuantityFormatter is registered with Spring MVC automatically. Added a couple of additional unit tests to verify the code can also parse slightly misprinted representations (mostly whitespace related).