#230 - New OrderCancelled event and overhaul of event handling in general.
We now publish an OrderCancelled event if an Order instance is cancelled, potentially roll back the inventory updates and create a compensating ProductPaymentEntry in the accountancy.

Fixed event handling in accountancy to handle OrderPaid events, not OrderCompleted ones. Polished implementation of InventoryOrderEventListener by tweaking the APIs in OrderCompletionReport for better composability.

Added Order.addOrderLine(…) as a replacement for ….add(OrderLine) and added lookup method to find OrderLine instances by product as the Order doesn't deduplicate them like the Cart does.

Added notes on the published events to state transitioning methods in OrderManager.