#232 - Catalog exposes methods to find products by multiple categories.
Added methods to Catalog to look up all products that have a set of categories assigned (i.e. they are assigned to all of the given categories) as well as the products that are assigned to any of given categories.

Fixed Product.toString() to use field access to avoid the appearance of Streamable.toString() in the output.