#234 - Reduce visibility of MonetaryAmountAttributeConverter.
Previously, this class needed to be public for the following reason: Hibernate (currently 5.3.7) insisted on re-scanning the configured persistence unit URL despite the fact that Spring already provides a completely configured persistence unit listing managed classes and packages. The result of that additional scanning attempt was immediately pre-processed and JPA AttributeConverter implementations found were handled by AttributeConverterDefinition which was deprecated in 5.3 as it – according to the deprecation notice – actually breaks the resolution of AttributeConverters via the managed bean infrastructure. This caused a raw reflection based instantiation attempt to MonetaryAmountAttributeConverter which then fails as the class is not made accessible beforehand (see [0] for my previous request to enable that).

However, [1] implements a Hibernate property that can be used to disable the scanning entirely, which we can do, as Boot's default configuration always provides a completely preconfigured persistence unit anyway. I've tweaked Salespoint bootstrap to always declare that property for now.

I've created a ticket in Spring Boot [2] to consider this flag to be set by default as the additional scanning shouldn't be neccesary in a default JPA Boot context as we already provide a fully prepared persistence unit to Hibernate.

[0] https://hibernate.atlassian.net/browse/HHH-10778
[1] https://hibernate.atlassian.net/browse/HHH-12505
[2] https://github.com/spring-projects/spring-boot/issues/15321