#163 - Relax Product to InventoryItem relationship.
We now offer two flavors of what's previously been known as Inventory. UniqueInventory retains the old semantics of a single (Unique)InventoryItem per Product and is the default choice for interacting and working with InventoryItems. In parallel to that, MultiInventoryItem has been introduced that allows a 1:n relationship between a Product and an InventoryItem. This can be used to model multiple warehouses that each keep track of their own stock.

The inventory updates upon order completion have been altered to only apply to UniqueInventoryItems. Also, guards are in place to prevent MultiInventoryItem being inserted alongside UniqueInventoryItem.

Reference documentation updated. Polished inventory module for var usage.