#243 - Fixed registration of LoggedInHandlerMethodArgumentResolver.
The fix for #241 removed the custom WebMvcConfigurer that explicitly registered all HandlerMethodArgumentResolvers available as Spring bean with Spring MVC. That class was removed as we assumed that Spring Boot would automatically register HMARs just as it does register Converters. Apparently it does not do that.

The issue went undetected as in the absence of LIHMAR, standard Spring MVC resolution simply instantiates an empty UserAccount and wraps that in an Optional. We now explicitly test for the availability of the LIHMAR within the RequestMappingHandlerAdapter.

Fixed line endings in LoggedInHandlerMethodArgumentResolver.

Related tickets: #241.