#269 - Tweak AbstractEntity.isNew in @PrePersist method instead of @PostPersist.
This is to make sure that the persistence provider actually flips the flag in CrudRepository.save(…) or EntityManager.persist(…) to be precise.

With the previous implementation the entity was returned from CrudRepository.save(…) as new which is wrong in itself but also caused test errors trying immediate subsequent deletions that run into an early return due to optimizations introduced for DATAJPA-1535.