#221 - Marked interfaces of injectable components with stereotype annotations.
As our implementation classes are package protected to avoid client code to refer to them, they don't appear in Salespoint's Javadoc. That makes it hard to identify which of the types are actually candidates for injection into student project's code.

We now explicitly annotate the public interfaces of those components with the Spring stereotype annotations although that sometimes wouldn't even be necessary (e.g. in the case of Spring Data repositories).