#280 - Roles are now properly propagated as role authorities to Spring Security.
We now automatically prefix Salespoint role names with ROLE_ when we prepare the granted authorities of the UserDetails instance we hand to Spring Security. This creates symmetry between creating the roles (via Role.of(…)) and actually using the role names in hasRole(…) expressions in the Thymeleaf integration and @PreAuthorize annotations. Previously, the ROLE_ prefix had to be used explicitly when creating the Role instances.