#300 - Properly load long-based Quantities from the database.
tl;dr: we now properly recreate Quantities created from long values when reading entities from the database.

Our domain model contains a Quantity type consisting of an amount (BigDecimal) and metric (an enum). One of the available metrics is UNIT which is usually used with non-decimal values like 2, 5 etc. Therefore, Quantity exposes ….of(…) methods taking a long argument forwarding that into a BigDecimal.valueOf(long) which creates a BigDecimal with precision of zero.

The Quantity type is a JPA @Embeddable so that Hibernate unfolds its properties into separate columns. Also, a QuantityAttributeConverter existed to properly recreate the long-bases Quantities and serializing the value into a String. While unit tests for those existed, the converter in fact was never used as apparently JPA does not apply converters to embeddables.

The problem that was now making its way into user applications was that Quantity.of(5, Metric.UNIT) would get the amount persisted using two decimal digits and thus return as an equivalent of Quantity.of(5.00, Metric.UNIT) which then in turn renders equality broken as BigDecimal 5 is not equal to BigDecimal 5.00.

What now follows is the trail of attempts to make sure a the long-based quantity values are returned properly:

1. Trying to get QuantityAttributeConverter work – I tried to remove @Embeddable from Quantity hoping that Hibernate would consider Quantity a basic type then and apply the converter as specified in JPA 2.2 section 3.8. What I got however was a

org.hibernate.MappingException: Could not determine type for: @Column(quantity)

2. Try to re-coerce the value into a long-based one in case the Metric is UNIT – ideally I'd just implement a constructor within Quantity inspecting both values and massaging the amount into a BigDecimal.valueOf(amount.longValue()) if the metric is UNIT. Unfortunately, JPA does not support complex constructors *at all*.

2a. Trying to use a lifecycle callback – With the lack of proper constructor support the next idea was a lifecycle callback (@PostLoad). Those, however, are only supported for entities (JPA 2.2 section 2.5.2)

2b. Using setters instead – my final (and sucessfull) attempt was to rather use the setter methods for the amount and metric attributes to apply the needed massaging with the following consequences:

- remove the final flag from both fields
- introducing package protected setters that apply the massaging in both directions as we don't know which of the methods is called first
- switching to property access
- moving our column customizations to the accessor methods
- explicitly declare some boolean is…(…) business methods as @Transient as otherwise Hibernate complains about the lack of setters.

This leaves us with an effectively final but way more cluttered variant of Quantity.