#306 - Fix Javadoc search.
We're now enforcing the build to run on JDK 13 as that has an issue in Javadoc fixed that breaks Javadoc search for non-module projects. Added Enforcer plugin to make sure we use a JDK 13. Switched Travis build to OpenJDK 13.

Added additional Javadoc references so that types from Spring Data and Spring Framework are linked properly.