#319 - Revamped build to allow different CI jobs.
Restructured the build to now contain a dedicated 'docs' profile that runs just enough tests to be able to render the Asciidoctor documentation. The default test run will exclude the documentation creation tests. This speeds up the standard build as it doesn't have to run that extra test and create the reference docs each and every time.

There are now two GitHub workflows:

- build.yaml - to package the artifact and deploy it into the Salespoint repository (on every commit to main)
- docs.yaml - triggers doc specific build and deploys the docs and the website to the server (nightly)

Tweaked the site upload to avoid attempting to chmod the uploaded files as it tries to do that on the root folder and the upload process hangs on that attempt.