#252 - More events in Inventory.
Decreasing the quantity of an InventoryItem now results in an internal QuantityReduced event being published. That is listened to by InternalInventoryListeners and verified against the threshold quanity defined via salespoint.inventory.restock-threshold (defaulting to Quantity.NONE). In case the current quantity is below that threshold a publicly consumable StockShort event is published that can be used to trigger further logic.

Slightly tweaked the arrangement in Inventory event listening. The event types are now member classes of InventoryEvents. Listeners are grouped into InventoryListeners and the actual business logic executed for the OrderCompleted and OrderCanceled events has been moved into an internal InventoryManagement.

Tweaked the Moduliths setup to always include the quantity module so that the Spring Boot type binding can be used in InventoryProperties. Had to let QuantityFormatter implement Converter as well and be annotated with @ConfigurationPropertiesBinding to be picked up.

AbstractEntity now exposes ….hasId(…) for easier comparison.