#304 - Prevent accidental usage of default constructor from aggregate subtypes.
We now explicitly verify the nullability of values that can only be null if object instances have been created through the JPA required default constructor. We avoid exposing those in the first place where possible (e.g. in UserAccount) as that class is not designed to be extended in the first place.

The verifications are implemented using @PrePersist hooks.