#364 - Integration test robustness.
Using Temurin JDK 17 on Ubunutu apparently uses a default Clock that uses nanoseconds precision for LocalDateTime. We have test cases that persiste those instances in the database and that defaults to only microseconds precision for timestamps. Thus, the tests comparing the persistent and the reference values fail on machines with such detailed precision.