#374 - SalespointRepository now extends PagingAndSortingRepository.
This transparently allows all repository implementations to be accessed with a provided Sort to define the ordering of the result. The plain ….findAll() doesn't actually even guarantee a stable order.

Redeclared all methods of PagingAndSortingRepository that return an Iterable to rather return Streamable for more convenient iteration, filtering and mapping.