#382 - Introduce abstraction for file storage.
Introduce the Storage API and an implementation based on a configured Path (defaulting to ~/.salespoint/uploads) to store files. The stored files can be obtained via Spring's Resource API so that they can easily be served for HTTP requests. Storing a file causes a FileStored event to be published for integration purposes.