#389 - @LoggedIn can now be used without Optional if guarded by authorization annotation.
We now do not reject the parameter value resolution of a plain UserAccount (not wrapped into an Optional) if the method declaring that parameter is guarded by Spring Security's @PreAuthorize annotation. Those annotations usually contain expressions that at least expect an authorization present and thus invocation will be aborted eventually with an authorization exception. The previous designed caused a ServletRequestBindingException during the preparation of the parameter and thus did not lead to Spring Security's default handling of unauthenticated access to resources that require authentication.

This also solves #337. We decided not to go with the Spring Security native integration for domain types in the context of @AuthenticationPrincipal as that would just cause different integration code needed to be written as we cannot store the UserAccount instance as principal in the authentication, as it is a managed JPA entity.