#365 - Adding up quantities of a CartItem to zero or negative removes item from Cart.
Changed CartItem to not allow quantities of zero or negative. Added Quantity.isPositive() and ….isZeroOrPositive() for symmetry with ….is(ZeroOr)Negative(). Operations updating the quantity for a product in the Cart now returns an Optional to represent that a calculated quantity of zero will remove the CartItem from the Cart and it might not be available anymore.