#366 - Cart now quantifies items and allows quantity lookup by product.
Quantity now has a ….toUnit() method that reduces non-unit quantities to a unit one with value 1. Cart uses that to quantify the elements contained in it using the actual units for unit-based product but resorting to the 1 for everything else. Cart now also exposes the Quantity for a given Product(Identifier).