GH-417 - Upgrade to Spring Boot 3.0 GA.
Remove the explicit downgrade of H2 to benefit from full H2 autoconfiguration (enablement of the console etc.). This requires us to enable Hibernate's automatic keyword quoting (via the `spring.jpa.properties.hibernate.auto_quote_keyword` property, fixes GH-414) as the new H2 version introduces quite a few new keywords that we use in our domain model (e.g. Interval.end). This in turn requires us to change the application of the default properties, we previously set in `Salespoint.RequiredPropertiesPostProcessor`. That is activated via spring.factories, which apparently is not considered when using ApplicationContextRunner. To avoid this, we now register the default properties via salespoint.properties in the classpath root and register that as property source on the Salespoint configuration class.

Also upgrade Spring Modulith to 0.1.0.