GH-423 - Allow accessing all UserAccounts for currently logged in users.
AuthenticationManagement now exposes getCurrentlyLoggedInUserAccounts() to obtain UserAccount instances for the users currently logged in.

To achieve that we now register a SpringSecurity SessionRegistry by default. UserAccountDetails now also keeps the UserAccountIdentifier around, so that the implementation of the method can access it directly. The identifiers could have been reified from the usernames already present in UAD, but that would've made the implementation dependent on the way that the UserAccountManagement implementation sets up the UserAccounts.