GH-423 - Polishing.
Reduce visibility of @Bean-methods in Salespoint configuration class. Reduce visibility of the class itself as from production code, it is only used within the very package anyway. Introduced TestSalespoint as public extension of Salespoint in the test sources to allow test cases to refer to the configuration class from other packages.