GH-305 - Support to change usernames.
We now use a simple UUID as backing value for the UserAccountIdentifier and additionally register a unique username as separate property. That allows code to freely change the username on a UserAccount assuming it stays unique within the application.

The UserAccount aggregate handling was changed to defensively detach all instances returned from UserAccountManagement. Without that, hanging a username on a managed entity effectively prevents proper application level error handling as issuing a query to potentially find conflicting other user accounts will prematurely flush the changed data, issue an update that in turn causes a DataIntegrityViolation exception.